requirements.txt
django==4.0.4
djangorestframework==3.13.1
dj-database-url==0.5.0
django-cors-headers==3.11.0
gunicorn==20.1.0
django-filter==21.1
psycopg2-binary==2.9.3
python-decouple==3.6